+++ src/KordanorsCabal.Game/Character/Descriptor/CharacterTypeDescriptor.vb
@@ -1,6 +1,6 @@
 ﻿Friend MustInherit Class CharacterTypeDescriptor
     MustOverride ReadOnly Property InitialStatistics As IReadOnlyDictionary(Of CharacterStatisticType, Long)
-    MustOverride ReadOnly Property MaximumEncumbrance(character As Character) As Single
+    MustOverride ReadOnly Property MaximumEncumbrance(character As Character) As Long
     MustOverride Function IsEnemy(character As Character) As Boolean
     MustOverride ReadOnly Property SpawnCount(level As DungeonLevel) As Long
     Overridable Function CanSpawn(location As Location, level As DungeonLevel) As Boolean
