+++ src/KordanorsCabal.Game/Character/Descriptor/Cult/PriestDescriptor.vb
@@ -20,7 +20,7 @@
         End Get
     End Property
 
-    Public Overrides ReadOnly Property MaximumEncumbrance(character As Character) As Single
+    Public Overrides ReadOnly Property MaximumEncumbrance(character As Character) As Long
         Get
             Return 0
         End Get
